ipn/ipnlocal: add helpers to manage ad-hoc locking during client state update
As a step toward normalizing the lock handling in SetControlClientStatus, push
the explicit unlock/lock pairs into "Locked" helpers. All these helpers do for
now is release and reacquire the lock, but that lets us push the special case
further down the call stack so we can simplify this big method.

Change-Id: Ife65c019eea675545991ae364366497f0be4c9d2
Signed-off-by: M. J. Fromberger <fromberger@tailscale.com>